+++ frontend/src/components/atoms/FormController.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
-import { Input, Form } from 'semantic-ui-react';
+import { Input, Form, Ref } from 'semantic-ui-react';
 import { Session } from 'model/index';
 /* eslint-disable react/jsx-props-no-spreading */
 
@@ -9,7 +9,8 @@ const FormController: FC<{
   label: string;
   icon: string;
   errormessage: string;
-}> = ({ name, label, icon, errormessage }) => {
+  required: boolean;
+}> = ({ name, label, icon, errormessage, required }) => {
   const {
     formState: { errors },
   } = useFormContext<Session>();
@@ -20,24 +21,27 @@ const FormController: FC<{
       rules={{
         required: errormessage,
       }}
-      render={({ field }) => (
-        <Form.Field
-          error={
-            errors[name] && {
-              content: errors[name]?.message,
-              pointing: 'below',
+      defaultValue=""
+      render={({ field: { ref, ...method } }) => (
+        <Ref innerRef={ref}>
+          <Form.Field
+            error={
+              errors[name] && {
+                content: errors[name]?.message,
+                pointing: 'below',
+              }
             }
-          }
-          control={Input}
-          label={label}
-          icon={icon}
-          type={name}
-          data-testid={name}
-          required
-          iconPosition="left"
-          placeholder={name}
-          {...field}
-        />
+            control={Input}
+            label={label}
+            icon={icon}
+            type={name}
+            data-testid={name}
+            required={required}
+            iconPosition="left"
+            placeholder={name}
+            {...method}
+          />
+        </Ref>
       )}
     />
   );
