+++ frontend/src/components/molecules/SignInForm.tsx
@@ -1,112 +1,41 @@
 import { FC } from 'react';
-import { Form, Grid, Segment, Input, Header } from 'semantic-ui-react';
-import { useForm, Controller } from 'react-hook-form';
-import { useHistory } from 'react-router-dom';
-import { Session } from 'model/index';
-import { Fetchsessionnew } from 'apis/Session';
-import { useSetRecoilState } from 'recoil';
-import LoginState from 'atom';
+import { Form, Grid, Segment, Header } from 'semantic-ui-react';
+import { FormProvider, UseFormReturn } from 'react-hook-form';
 import FormMessage from 'components/atoms/FormMessage';
+import FormController from 'components/atoms/FormController';
+import { Session } from 'model/index';
+/* eslint-disable react/jsx-props-no-spreading */
 
-const SignInForm: FC = () => {
-  const history = useHistory();
-  const {
-    control,
-    formState: { errors },
-    handleSubmit,
-  } = useForm<Session>({ criteriaMode: 'all' });
-
-  const setUser = useSetRecoilState(LoginState);
-
-  const onSubmit = async (data: Session) => {
-    await Fetchsessionnew(data)
-      .then((result) =>
-        result !== undefined && result.data
-          ? (setUser((prevUser) => ({ ...prevUser, ...result.data })),
-            history.push('/', {
-              message: 'ログインに成功しました。',
-              type: 'success',
-            }))
-          : history.push('/sign_in', {
-              message: 'ログインに失敗しました。',
-              type: 'error',
-            }),
-      )
-      .catch(() =>
-        history.push('/sign_in', {
-          message: 'エラーが発生しました。',
-          type: 'error',
-        }),
-      );
-  };
-
-  return (
-    <>
-      <Header
-        as="h1"
-        content="ログイン"
-        textAlign="center"
-        style={{ marginBottom: '1rem' }}
-      />
-      <Form onSubmit={handleSubmit(onSubmit)}>
+const SignInForm: FC<{
+  methods: UseFormReturn<Session>;
+  onSubmit: (data: Session) => Promise<void>;
+}> = ({ methods, onSubmit }) => (
+  <>
+    <Header
+      as="h1"
+      content="ログイン"
+      textAlign="center"
+      style={{ marginBottom: '1rem' }}
+    />
+    <FormProvider {...methods}>
+      <Form onSubmit={methods.handleSubmit(onSubmit)}>
         <Grid columns={3} centered style={{ margin: '4em' }}>
           <Grid.Column width={3} />
           <Grid.Column width={10}>
             <Segment>
-              <Controller
+              <FormController
                 name="email"
-                control={control}
-                rules={{
-                  required: 'メールアドレスが入力されていません。',
-                }}
-                render={({ field: { onChange, onBlur, value } }) => (
-                  <Form.Field
-                    error={
-                      errors.email && {
-                        content: errors.email?.message,
-                        pointing: 'below',
-                      }
-                    }
-                    control={Input}
-                    label="メールアドレス"
-                    icon="mail"
-                    data-testid="email"
-                    required
-                    iconPosition="left"
-                    placeholder="e-mail"
-                    onChange={onChange}
-                    onBlur={onBlur}
-                    value={value}
-                  />
-                )}
+                label="メールアドレス"
+                icon="mail"
+                errormessage="メールアドレスが入力されていません。"
+                required
               />
-              <Controller
+              <FormController
                 name="password"
-                control={control}
-                rules={{
-                  required: 'パスワードが入力されていません。',
-                }}
-                render={({ field: { onChange, onBlur, value } }) => (
-                  <Form.Field
-                    error={
-                      errors.password && {
-                        content: errors.password?.message,
-                        pointing: 'below',
-                      }
-                    }
-                    control={Input}
-                    label="パスワード"
-                    icon="key"
-                    data-testid="password"
-                    required
-                    type="password"
-                    iconPosition="left"
-                    placeholder="password"
-                    onChange={onChange}
-                    onBlur={onBlur}
-                    value={value}
-                  />
-                )}
+                label="パスワード"
+                icon="key"
+                errormessage="パスワードが入力されていません。"
+                required
               />
               <Form.Field
                 style={{ textAlign: 'center', justifyContent: 'center' }}
@@ -123,8 +52,8 @@ const SignInForm: FC = () => {
           <Grid.Column width={3} />
         </Grid>
       </Form>
-    </>
-  );
-};
+    </FormProvider>
+  </>
+);
 
 export default SignInForm;
