+++ frontend/src/components/pages/Session.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Grid } from 'semantic-ui-react';
 import { Helmet } from 'react-helmet';
-import SignInForm from 'container/EnhancedSignIn';
+import SignInForm from 'container/EnhancedSignInForm';
 import SignOut from 'container/EnhancedSignOut';
 import Confirmation from 'components/molecules/Confirmation';
 import ResetPassword from 'components/organisms/ResetPassword';
