+++ frontend/src/container/EnhancedSignInForm.tsx
@@ -1,0 +1,12 @@
+import { FC } from 'react';
+import SignInForm from 'components/molecules/SignInForm';
+import useSignInForm from 'hooks/SignInForm';
+/* eslint-disable react/jsx-props-no-spreading */
+
+const EnhancedSignInForm: FC = () => {
+  const form = useSignInForm();
+
+  return <SignInForm {...form} />;
+};
+
+export default EnhancedSignInForm;
